+++ CHANGELOG.md
@@ -17,3 +17,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `StringEventChannel` variant
 - `TransformEventChannel` variant
 - `VoidEventChannel` variant
+- Channels sample scene